Back_taller/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pymongo==4.5.0
dnspython==2.4.2
PyJWT==2.8.0
gunicorn==21.2.0
bcrypt==4.0.1
email-validator==2.0.0
marshmallow==3.20.1
Flask-Limiter==3.5.0 